# swagger_full.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "App Base API",
        "description": "API REST base com autenticação JWT, gestão de utilizadores, upload de ficheiros e sistema de refresh tokens. Desenvolvida para servir como ponto de partida em novos projectos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-parceiro-dexa.appteste.info"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
...
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Listar os meus tickets",
                "description": "Retorna todos os tickets publicados pelo parceiro autenticado (paginado)",
                "operationId": "listar_1",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "tamanho",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de tickets",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseTicketResumoResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado — requer papel PARCEIRO",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Submeter um ticket",
                "description": "Cria um novo ticket de suporte com estado ABERTO",
                "operationId": "criar_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TicketRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Ticket criado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado — requer papel PARCEIRO",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/tickets/{publicId}/estado": {
            "patch": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Alterar estado do ticket (Admin)",
                "description": "Actualiza o estado de um ticket específico (ABERTO, PROGRESSO, RESOLVIDO). Acesso exclusivo ao ADMIN.",
                "operationId": "alterarEstado",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TicketEstadoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Estado actualizado com sucesso"
                    },
                    "400": {
                        "description": "Estado inválido",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado — requer papel ADMIN",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/tickets/{publicId}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Detalhes do ticket",
                "description": "Retorna todas as informações de um ticket específico",
                "operationId": "buscarDetalhes_2",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Eliminar ticket",
                "description": "Elimina permanentemente um ticket. O ADMIN pode eliminar qualquer ticket; o PARCEIRO só pode eliminar os seus próprios.",
                "operationId": "eliminar_4",
                "parameters": [
                    {
                        "name": "publicId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ticket eliminado com sucesso"
                    },
                    "400": {
                        "description": "Sem permissão para eliminar o ticket de outro parceiro",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StandardError": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "msg": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "TicketRequest": {
                "type": "object",
                "properties": {
                    "conteudo": {
                        "type": "string",
                        "minLength": 1
                    },
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "CANCELAR",
                            "BUG",
                            "RECLAMACAO",
                            "OUTRO"
                        ]
                    }
                },
                "required": [
                    "conteudo",
                    "tipo"
                ]
            },
...
            "TicketEstadoRequest": {
                "type": "object",
                "properties": {
                    "estado": {
                        "type": "string",
                        "enum": [
                            "ABERTO",
                            "PROGRESSO",
                            "RESOLVIDO"
                        ]
                    }
                },
                "required": [
                    "estado"
                ]
            },
...
            "PageResponseTicketResumoResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TicketResumoResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
            "TicketResumoResponse": {
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string"
                    },
                    "conteudo": {
                        "type": "string"
                    },
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "CANCELAR",
                            "BUG",
                            "RECLAMACAO",
                            "OUTRO"
                        ]
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "ABERTO",
                            "PROGRESSO",
                            "RESOLVIDO"
                        ]
                    },
                    "publicadoPorNome": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    }
                }
            },
            "TicketResponse": {
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string"
                    },
                    "conteudo": {
                        "type": "string"
                    },
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "CANCELAR",
                            "BUG",
                            "RECLAMACAO",
                            "OUTRO"
                        ]
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "ABERTO",
                            "PROGRESSO",
                            "RESOLVIDO"
                        ]
                    },
                    "publicadoPorPublicId": {
                        "type": "string"
                    },
                    "publicadoPorNome": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "name": "Bearer Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}